+++ workspaces/bulk-import/plugins/bulk-import/src/types/types.ts
@@ -119,6 +119,7 @@ export enum RepositorySelection {
 export enum ApprovalTool {
   Git = 'git',
   ServiceNow = 'servicenow',
+  Gitlab = 'gitlab',
 }
 
 export type CreateImportJobRepository = {