+++ workspaces/lightspeed/plugins/lightspeed-backend/src/service/router.ts
@@ -32,7 +32,6 @@ import {
 } from '@red-hat-developer-hub/backstage-plugin-lightspeed-common';
 
 import { userPermissionAuthorization } from './permission';
-import { QUERY_SYSTEM_INSTRUCTION } from './prompt';
 import {
   DEFAULT_HISTORY_LENGTH,
   QueryRequestBody,
@@ -158,10 +157,7 @@ export async function createRouter(
     '/v1/query',
     validateCompletionsRequest,
     async (request, response) => {
-      const {
-        provider,
-        system_prompt,
-      }: Pick<QueryRequestBody, 'provider' | 'system_prompt'> = request.body;
+      const { provider }: Pick<QueryRequestBody, 'provider'> = request.body;
       try {
         const credentials = await httpAuth.credentials(request);
         const userEntity = await userInfo.getUserInfo(credentials);
@@ -174,9 +170,6 @@ export async function createRouter(
           credentials,
         );
         const userQueryParam = `user_id=${encodeURIComponent(user_id)}`;
-        if (!system_prompt) {
-          request.body.system_prompt = QUERY_SYSTEM_INSTRUCTION;
-        }
         request.body.media_type = 'application/json'; // set media_type to receive start and end event
         const requestBody = JSON.stringify(request.body);
         const fetchResponse = await fetch(