+++ workspaces/marketplace/plugins/marketplace/src/components/MarketplacePluginInstallContent.tsx
@@ -32,6 +32,7 @@ import {
   MarketplacePackageSpecAppConfigExample,
   MarketplacePlugin,
 } from '@red-hat-developer-hub/backstage-plugin-marketplace-common';
+
 import { JsonObject } from '@backstage/types';
 
 import Box from '@mui/material/Box';
@@ -47,6 +48,7 @@ import FormControlLabel from '@mui/material/FormControlLabel';
 import Checkbox from '@mui/material/Checkbox';
 import Button from '@mui/material/Button';
 import Typography from '@mui/material/Typography';
+import Tooltip from '@mui/material/Tooltip';
 
 import { pluginInstallRouteRef, pluginRouteRef } from '../routes';
 import { usePlugin } from '../hooks/usePlugin';
@@ -59,6 +61,7 @@ import {
   useCodeEditor,
 } from './CodeEditor';
 import { Markdown } from './Markdown';
+import { useExtensionConfigPermission } from '../hooks/useExtensionConfigPermission';
 
 const generateCheckboxList = (packages: MarketplacePackage[]) => {
   const hasFrontend = packages.some(
@@ -218,6 +221,10 @@ export const MarketplacePluginInstallContent = ({
   }, [codeEditor, packages]);
 
   const navigate = useNavigate();
+  const canInstallPlugin = useExtensionConfigPermission(
+    params.namespace,
+    params.name,
+  );
   const examples = packages[0]?.spec?.appConfigExamples;
   const installationInstructions = plugin.spec?.installation;
   const aboutMarkdown = plugin.spec?.description;
@@ -395,9 +402,19 @@ export const MarketplacePluginInstallContent = ({
         <Box sx={{ mt: 1, mb: 2, display: 'none' }}>
           <CheckboxList packages={packages} />
         </Box>
-        <Button variant="contained" color="primary" disabled>
-          Install
-        </Button>
+        <Tooltip
+          title={
+            !canInstallPlugin.data?.authorizedActions?.includes('create')
+              ? "You don't have permission to install plugins or edit their configurations. Contact your administrator to request access or assistance."
+              : ''
+          }
+        >
+          <Typography component="span">
+            <Button variant="contained" color="primary" disabled>
+              Install
+            </Button>
+          </Typography>
+        </Tooltip>
         <Button
           variant="outlined"
           color="primary"
@@ -406,14 +423,16 @@ export const MarketplacePluginInstallContent = ({
         >
           Cancel
         </Button>
-        <Button
-          variant="text"
-          color="primary"
-          onClick={onLoaded}
-          sx={{ ml: 3 }}
-        >
-          Reset
-        </Button>
+        {canInstallPlugin.data?.authorizedActions?.includes('create') && (
+          <Button
+            variant="text"
+            color="primary"
+            onClick={onLoaded}
+            sx={{ ml: 3 }}
+          >
+            Reset
+          </Button>
+        )}
       </Box>
     </Box>
   );