+++ workspaces/orchestrator/plugins/orchestrator-form-widgets/http-workflow-dev-server/index.js
@@ -27,6 +27,7 @@ const logRequest = req => {
   console.log('request: ', {
     originalUrl: req.originalUrl,
     method: req.method,
+    query: req.query,
     headers: req.headers,
     body: req.body,
   });
@@ -116,9 +117,52 @@ app.post('/chunk02', (req, res) => {
   res.send(JSON.stringify(chunk02));
 });
 
+app.get('/activeTextWhisperer', (req, res) => {
+  logRequest(req);
+
+  const mydata = req.query.mydata;
+
+  const autocomplete = ['my option one', 'my option two', 'Jack', 'Joe'];
+  if (mydata && mydata !== '___undefined___') {
+    autocomplete.push(mydata);
+  }
+
+  const result = {
+    myresult: { foo: { default: 'This is dynamically fetched default value' } },
+    bar: { something: { myautocompleteoptions: autocomplete } },
+  };
+
+  res.send(JSON.stringify(result));
+});
+
 app.listen(port, () => {
   // eslint-disable-next-line no-console
   console.info(
     `Simple HTTP server for orchestrator-form-widgets development only. Listening on ${port} port.`,
   );
+});
+
+app.post('/validate', (req, res) => {
+  logRequest(req);
+
+  const field = req.body?.field;
+  const value = req.body?.value;
+  const moreDataForMyValidator = req.body?.moreDataForMyValidator;
+
+  if (
+    field === 'mySimpleActiveText' &&
+    moreDataForMyValidator !== 'ignoreerror'
+  ) {
+    if (!value || value.length < 5) {
+      res.status(422);
+      res.send({
+        [field]: ['The field must be 5 or more characters long.'],
+      });
+
+      return;
+    }
+  }
+
+  res.status(200);
+  res.send('Valid');
 });