+++ workspaces/orchestrator/plugins/orchestrator-form-widgets/src/widgets/SchemaUpdater.tsx
@@ -13,31 +13,26 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import React, { useEffect, useMemo, useState } from 'react';
+import React, { useMemo, useState } from 'react';
 import { Widget } from '@rjsf/utils';
 import { JSONSchema7 } from 'json-schema';
 import { JsonObject } from '@backstage/types';
-import { makeStyles, Theme } from '@material-ui/core';
 import {
   useWrapperFormPropsContext,
   SchemaChunksResponse,
 } from '@red-hat-developer-hub/backstage-plugin-orchestrator-form-api';
 import { fetchApiRef, useApi } from '@backstage/core-plugin-api';
+import { useDebounce } from 'react-use';
 
 import { FormContextData } from '../types';
 import {
-  evaluateTemplate,
-  getRequestInit,
+  useRequestInit,
+  useEvaluateTemplate,
   useRetriggerEvaluate,
   useTemplateUnitEvaluator,
 } from '../utils';
-import { isEqual } from 'lodash';
-
-const useStyles = makeStyles((theme: Theme) => ({
-  error: {
-    color: theme.palette.error.main,
-  },
-}));
+import { ErrorText } from './ErrorText';
+import { DEFAULT_DEBOUNCE_LIMIT } from './constants';
 
 export const SchemaUpdater: Widget<
   JsonObject,
@@ -47,13 +42,8 @@ export const SchemaUpdater: Widget<
   const fetchApi = useApi(fetchApiRef);
   const templateUnitEvaluator = useTemplateUnitEvaluator();
 
-  const classes = useStyles();
   const formContext = useWrapperFormPropsContext();
-  const [_, setLoading] = useState(true);
   const [error, setError] = useState<string>();
-  const [evaluatedFetchUrl, setEvaluatedFetchUrl] = useState<string>();
-  const [evaluatedRequestInit, setEvaluatedRequestInit] =
-    useState<RequestInit>();
 
   const { updateSchema, formData } = formContext;
 
@@ -70,85 +60,79 @@ export const SchemaUpdater: Widget<
     uiProps['fetch:retrigger'] as string[],
   );
 
-  useEffect(() => {
-    evaluateTemplate({
-      template: fetchUrl,
-      key: 'fetch:url',
-      unitEvaluator: templateUnitEvaluator,
-      formData,
-    })
-      .then(evaluated => setEvaluatedFetchUrl(evaluated))
-      .catch(reason => setError(reason.toString()));
-  }, [fetchUrl, templateUnitEvaluator, formData]);
+  const evaluatedFetchUrl = useEvaluateTemplate({
+    template: fetchUrl,
+    key: 'fetch:url',
+    formData,
+    setError,
+  });
+  const evaluatedRequestInit = useRequestInit({
+    uiProps,
+    prefix: 'fetch',
+    formData,
+    setError,
+  });
 
-  useEffect(() => {
-    getRequestInit(uiProps, 'fetch', templateUnitEvaluator, formData)
-      .then(evaluated =>
-        setEvaluatedRequestInit(actual =>
-          isEqual(actual, evaluated) ? actual : evaluated,
-        ),
-      )
-      .catch(reason => setError(reason.toString()));
-  }, [uiProps, templateUnitEvaluator, formData]);
-
-  useEffect(() => {
-    const fetchSchemaChunks = async () => {
+  useDebounce(
+    () => {
       if (!evaluatedFetchUrl || !retrigger || !evaluatedRequestInit) {
         return;
       }
 
-      try {
-        setLoading(true);
-        setError(undefined);
+      const fetchSchemaChunks = async () => {
+        try {
+          setError(undefined);
 
-        const response = await fetchApi.fetch(
-          evaluatedFetchUrl,
-          evaluatedRequestInit,
-        );
-        const data = (await response.json()) as unknown as SchemaChunksResponse;
+          const response = await fetchApi.fetch(
+            evaluatedFetchUrl,
+            evaluatedRequestInit,
+          );
+          const data =
+            (await response.json()) as unknown as SchemaChunksResponse;
 
-        // validate received response before updating
-        if (!data) {
-          throw new Error('Empty response received');
-        }
-        if (typeof data !== 'object') {
-          throw new Error('JSON object expected');
-        }
-        Object.keys(data).forEach(key => {
-          if (!data[key].type) {
-            throw new Error(
-              `JSON response malformed, missing "type" field for "${key}" key`,
-            );
+          // validate received response before updating
+          if (!data) {
+            throw new Error('Empty response received');
           }
-        });
+          if (typeof data !== 'object') {
+            throw new Error('JSON object expected');
+          }
+          Object.keys(data).forEach(key => {
+            if (!data[key].type) {
+              throw new Error(
+                `JSON response malformed, missing "type" field for "${key}" key`,
+              );
+            }
+          });
 
-        updateSchema(data);
-      } catch (err) {
-        // eslint-disable-next-line no-console
-        console.error(
-          'Error when updating schema',
-          props.id,
-          evaluatedFetchUrl,
-          err,
-        );
-        setError(
-          `Failed to fetch schema update by the ${props.id} SchemaUpdater`,
-        );
-      } finally {
-        setLoading(false);
-      }
-    };
+          updateSchema(data);
+        } catch (err) {
+          // eslint-disable-next-line no-console
+          console.error(
+            'Error when updating schema',
+            props.id,
+            evaluatedFetchUrl,
+            err,
+          );
+          setError(
+            `Failed to fetch schema update by the ${props.id} SchemaUpdater`,
+          );
+        }
+      };
 
-    fetchSchemaChunks();
-  }, [
-    evaluatedFetchUrl,
-    evaluatedRequestInit,
-    fetchApi,
-    props.id,
-    updateSchema,
-    // no need to expand the "retrigger" array here since its identity changes only if an item changes
-    retrigger,
-  ]);
+      fetchSchemaChunks();
+    },
+    DEFAULT_DEBOUNCE_LIMIT,
+    [
+      evaluatedFetchUrl,
+      evaluatedRequestInit,
+      fetchApi,
+      props.id,
+      updateSchema,
+      // no need to expand the "retrigger" array here since its identity changes only if an item changes
+      retrigger,
+    ],
+  );
 
   if (!fetchUrl) {
     // eslint-disable-next-line no-console
@@ -161,7 +145,7 @@ export const SchemaUpdater: Widget<
   }
 
   if (error) {
-    return <div className={classes.error}>{error}</div>;
+    return <ErrorText text={error} />;
   }
 
   // No need to render anything