+++ workspaces/sandbox/plugins/sandbox/src/api/RegistrationBackendClient.tsx
@@ -141,7 +141,14 @@ export class RegistrationBackendClient implements RegistrationService {
 
     if (!response.ok) {
       const error: CommonResponse = await response.json();
-      throw new Error(error?.message);
+      // handle backend error messages and make it more user-friendly
+      if (error?.message.includes("Invalid 'To' Phone Number")) {
+        throw new Error(
+          'Invalid phone number. Please verify the country code and number format, then try again.',
+        );
+      } else {
+        throw new Error(error?.message);
+      }
     }
   };